comment userId 추가

+++ clonecoding/src/main/java/project7/clonecoding/comment/dto/CommentResponseDto.java
@@ -9,5 +9,5 @@ public class CommentResponseDto {
     private String userName;
     private String createdDate;
     private String comment;
-    private int star;
+    private Float star;
 }
